requirements.txt
Flask==2.3.0
Flask-Login==0.6.2
SQLAlchemy==2.0.15
Werkzeug==2.3.0
Pillow==10.0.0
requests==2.31.0
replicate==0.15.4
